Move battery specific code to _battery_manager.py
This just moves the code out, so that it becomes easier to generalize
and cleanup.
