+++ src/frequenz/sdk/actor/power_distributing/_component_managers/__init__.py
@@ -1,0 +1,12 @@
+# License: MIT
+# Copyright © 2023 Frequenz Energy-as-a-Service GmbH
+
+"""Interfaces for the power distibuting actor with different component types."""
+
+from ._battery_manager import BatteryManager
+from ._component_manager import ComponentManager
+
+__all__ = [
+    "BatteryManager",
+    "ComponentManager",
+]
