+++ src/frequenz/sdk/timeseries/battery_pool/_metric_calculator.py
@@ -11,11 +11,11 @@ from datetime import datetime, timezone
 from typing import Generic, TypeVar
 
 from ... import timeseries
+from ...actor.power_distributing._component_managers._battery_manager import (
+    _get_battery_inverter_mappings,
+)
 from ...actor.power_distributing._distribution_algorithm._distribution_algorithm import (
     _aggregate_battery_power_bounds,
-)
-from ...actor.power_distributing.power_distributing import (
-    _get_battery_inverter_mappings,
 )
 from ...actor.power_distributing.result import PowerBounds
 from ...microgrid.component import ComponentMetricId
